requirements.txt
discord.py
python-dotenv
google-generativeai>=0.8.3
pypdf
aiohttp